Restructure app framework to facilitate api versioning
