+++ api/v1/controllers/ordersController.js
@@ -1,0 +1,10 @@
+export default class {
+  constructor(user) {
+    this.user = user;
+  }
+
+  index() {
+    // controller returns an empty object
+    return new this.Promise(resolve => resolve({}));
+  }
+}
